.NET Core 3 Upgrade (#430)
### ASP.NET Core API Boxed
- Upgrade to ASP.NET Core 3.0.
- Switch to use the generic host builder.
- Upgrade to Serilog.AspNetCore 3.2.0 and it's middleware based request logging.
- Change `Startup` to implement `StartupBase` instead of `IStartup`.
- Change paging to use a GraphQL like approach.
- Add lots more tests.
- Change integration tests to run all tests in parallel instead of relying on xUnit's `IClassFixture`.
- Add `TestStartup` class in integration tests instead of relying on the flaky extension methods on the test host.
- Add documented 500 ProblemDetails response to all API endpoints.
- Add more startup configurations to `launchSettings.json`.
- Change to only write logs to the console in development.
- Update .editorconfig file.
- Update .gitattributes file.
- Upgrade Microsoft.AspNetCore.Mvc.Versioning from `2.3.0` to `3.2.1`.
- Fix passing cancellation tokens in controller.
- Remove `IsPackable` from integration test because xUnit includes it.
- Get `IOptions` from `IConfiguration` instead of `BuildServiceProvider` and then `Resolve`.
- Use `is null` instead of `== null`.

### ASP.NET Core GraphQL Boxed
- Upgrade to ASP.NET Core 3.0.
- Switch to use the generic host builder.
- Upgrade to Serilog.AspNetCore 3.2.0 and it's middleware based request logging.
- Change `Startup` to implement `StartupBase` instead of `IStartup`.
- Change integration tests to run all tests in parallel instead of relying on xUnit's `IClassFixture`.
- Add `TestStartup` class in integration tests instead of relying on the flaky extension methods on the test host.
- Add more startup configurations to `launchSettings.json`.
- Change to only write logs to the console in development.
- Update .editorconfig file.
- Update .gitattributes file.
- Remove `IsPackable` from integration test because xUnit includes it.
- Get `IOptions` from `IConfiguration` instead of `BuildServiceProvider` and then `Resolve`.
- Use `is null` instead of `== null`.

### Microsoft Orleans Boxed
- Upgrade to .NET Core 3.0.
- Upgrade to Orleans `3.0.1`.
- Upgrade to Serilog 2.9.0.
- Switch to use the generic host builder.
- Add more startup configurations to `launchSettings.json`.
- Update .editorconfig file.
- Update .gitattributes file.
- Fix calling Dispose on integration test cluster.
- Remove `IsPackable` from integration test because xUnit includes it.
- Get `IOptions` from `IConfiguration` insteaof `BuildServiceProvider` and then `Resolve`.
- Use `is null` instead of `== null`.

+++ build.cake
@@ -63,23 +63,30 @@ Task("InstallDeveloperCertificate")
         if (isDotnetRunEnabled)
         {
             var certificateFilePath = System.IO.Path.ChangeExtension(System.IO.Path.GetTempFileName(), ".pfx");
-            StartProcess(
-                "dotnet",
-                new ProcessArgumentBuilder()
-                    .Append("dev-certs")
-                    .Append("https")
-                    .AppendSwitch("--export-path", certificateFilePath));
-            Information($"Dotnet Developer Certificate saved");
-
-            var certificate = new X509Certificate2(certificateFilePath);
-            using (var store = new X509Store(StoreName.Root, StoreLocation.LocalMachine))
+            try
             {
-                store.Open(OpenFlags.ReadWrite);
-                store.Add(certificate);
-            }
-            Information($"Dotnet Developer Certificate installed to local machine");
+                StartProcess(
+                    "dotnet",
+                    new ProcessArgumentBuilder()
+                        .Append("dev-certs")
+                        .Append("https")
+                        .AppendSwitch("--export-path", certificateFilePath));
 
-            System.IO.File.Delete(certificateFilePath);
+                var certificate = new X509Certificate2(certificateFilePath);
+                using (var store = new X509Store(StoreName.Root, StoreLocation.LocalMachine))
+                {
+                    store.Open(OpenFlags.ReadWrite);
+                    store.Add(certificate);
+                }
+                Information($"Dotnet developer certificate installed to local machine's root certificates.");
+            }
+            finally
+            {
+                if (System.IO.File.Exists(certificateFilePath))
+                {
+                    System.IO.File.Delete(certificateFilePath);
+                }
+            }
         }
         else
         {
@@ -101,7 +108,7 @@ Task("Test")
                 NoBuild = true,
                 NoRestore = true,
                 ResultsDirectory = artefactsDirectory,
-                // ArgumentCustomization = x => x.Append($"--logger html;LogFileName={project.GetFilenameWithoutExtension()}.html"),
+                ArgumentCustomization = x => x.Append($"--logger html;LogFileName={project.GetFilenameWithoutExtension()}.html"),
             });
     });
